attempt at implementing background
the background overlaps all the elements in page.tsx. tried ot to set the children on the layout into its own seperate div, aswell as the background and clock in the body itself.  might not be a good solution.

(background currently commented out on layout.tsx, uncomment to see problem).