Backend overhaul
1. lib/employeeStorage created: This file has a central employeeList to keep track of all employees, with different functions to fetch, update, clockIn or out etc.
for example: updateEmployeeStatus(), getPresentEmployees() etc can be used in the frontend components; backend abstracted away from frontend components

2. lib/dataAcess.ts and api routes updated: Data is fetched and formatted to respective employee objects in dataacess, and organized in employeeStorage.ts. the api routes have been optimized

In the data acess class it will use methods from employeeStorage to update status, forcing a refresh of the employeeList;

3. Definitions updated: employee is now the central type, and has an optional status for  isClockedIn? shiftStart? and shiftEnd?

this makes it easier to initalize employees and then we update as needed

The frontend components have not been changed. Need overhaul of how data is fetched into the frontend components. Use EmployeeStorage.ts for different functions of fetching data in an organized way