Overtime functionality
Moved the workTimeData that fetches and sums up this weeks worth of work-intervals to a context. The context can be used to now calculate fleks salary before clocking in.

In performCheckout overtime is now calculated with moment.js, formatted to a interval string that fits the database, and posted to the database.

Added imports for fonts on css. On vercel the fonts arent being loaded, but they are locally.