Notifications & Auto-close page
Added the "react-toastify" library, which makes it easy to implement smooth notifications. Implemented a success / error notification on clock operation.

Added ./hooks/useAutoClosePage which sets timers for closing the page
automatically, and when to display a toast. Resets and clears toast when activity is detected (mouse+touch event).

Fixed error in clockHistory/index.tsx causing vercel to fail deployment.