Context & dataflow rework
Main change is instead of getting the employee in ./components/employeePageData/ and then passing it down to every single component, then passing it down again and again to its children; theres now one context for it, selectedEmployeeContext.tsx
This way we can just get the employee in question without having to pass it through every component. When you go to the employee page, it finds the employeeId from the query and maps over the employees array to set the selectedEmployee. This is so much less clunky.

Implementing the rest of the context insteaad of passing down employee object in the employeePageData is still WIP, though.

Also: Instead of one context for timer, and one for posting and fetching the balance (formerly workTimeData), now its under one central context. This context is wrapped around the subcomponents of the employee page.